requirements.txt
# DeepBuild - Autonomous Coding Agent
# Python Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# LLM Integration
ollama>=0.1.0

# Web Search (support tool)
duckduckgo-search>=4.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
